+++ src/my_cave.hpp
@@ -16,9 +16,9 @@ typedef struct {
   //
   // Keeps track of the largest blob so fat
   //
-  uint8_t  largest_at_x;
-  uint8_t  largest_at_y;
-  uint16_t largest_size;
+  uint8_t largest_at_x;
+  uint8_t largest_at_y;
+  int     largest_size;
 
   //
   // Unique id per blob
@@ -42,7 +42,7 @@ typedef struct {
 } Cave;
 
 void cave_dump(Gamep g, Cave * /*c*/);
-void cave_create(Gamep g, Cave * /*c*/, uint32_t fill_prob, int r1, int r2, int map_generations);
+void cave_create(Gamep g, Cave * /*c*/, uint32_t fill_prob, uint8_t r1, uint8_t r2, int map_generations);
 int  cave_generation_fill_blob_cand(Gamep g, Cave *c, int x, int y, uint16_t size, uint16_t id);
 void cave_generation_keep_largest_blob(Gamep g, Cave *c);
 void cave_generation_center_blob(Gamep g, Cave *c);
